requirements.txt
aiogram==2.19
aiohttp==3.8.1
aioschedule==0.5.2
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
backports.entry-points-selectable==1.1.0
CacheControl==0.12.6
cachy==0.3.0
certifi==2021.10.8
cffi==1.14.6
cleo==0.8.1
clikit==0.6.2
colorama==0.4.4
crashtest==0.3.1
cryptography==3.4.8
distlib==0.3.2
filelock==3.0.12
frozenlist==1.3.0
html5lib==1.1
jeepney==0.7.1
keyring==21.8.0
lockfile==0.12.2
loguru==0.6.0
msgpack==1.0.2
multidict==6.0.2
packaging==20.9
pastel==0.2.1
peewee==3.14.10
pexpect==4.8.0
pipenv==2021.5.29
pkginfo==1.7.1
platformdirs==2.3.0
poetry==1.1.8
poetry-core==1.0.4
ptyprocess==0.7.0
pyaes==1.6.1
pyasn1==0.4.8
pycparser==2.20
pylev==1.4.0
pyparsing==2.4.7
python-dotenv==0.19.2
pytz==2022.1
pywin32-ctypes==0.2.0
requests==2.26.0
requests-toolbelt==0.9.1
rsa==4.8
SecretStorage==3.3.1
shellingham==1.4.0
six==1.16.0
Telethon==1.24.0
tomlkit==0.7.2
urllib3==1.26.6
virtualenv==20.7.2
virtualenv-clone==0.5.6
webencodings==0.5.1
win32-setctime==1.1.0
yarl==1.7.2
